+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @svarta/cli
 
+## 0.0.10
+
+### Patch Changes
+
+- run middlewares before input validation
+- Updated dependencies
+  - @svarta/adapter-standalone@0.0.10
+  - @svarta/dev-server@0.0.10
+  - @svarta/core@0.0.10
+
 ## 0.0.9
 
 ### Patch Changes